Alright supplying natal kicks works now, also passing opening angle and bhsigmafrac works. The default is to have an opening angle of 90.0, Mike was right the default of 0 which seemed to be the case in COSMIC right now meant that the kick could only happen at the poles. Also the bhsigmafrac is defaulting one meaning that BH get same kick as NS.

+++ cosmic/src/const_bse.h
@@ -27,8 +27,8 @@
       COMMON /POINTS/ pts1,pts2,pts3
       REAL*8 dmmax,drmax
       COMMON /TSTEPC/ dmmax,drmax
-      REAL scm(50000,14),spp(20,3)
+      REAL*8 scm(50000,14),spp(20,3)
       COMMON /SINGLE/ scm,spp
-      REAL bcm(50000,42),bpp(80,14)
+      REAL*8 bcm(50000,42),bpp(80,14)
       COMMON /BINARY/ bcm,bpp
 *
